tests/requirements.txt
docker
PyGithub
numpy